+++ projects/workforce-allocations-to-delius/src/main/kotlin/uk/gov/justice/digital/hmpps/integrations/delius/audit/BusinessInteractionCode.kt
@@ -1,4 +1,3 @@
 package uk.gov.justice.digital.hmpps.integrations.delius.audit
 
-enum class BusinessInteractionCode(val code: String) {
-}
+enum class BusinessInteractionCode(val code: String)
