+++ projects/workforce-allocations-to-delius/src/main/kotlin/uk/gov/justice/digital/hmpps/integrations/delius/contact/Contact.kt
@@ -74,7 +74,7 @@ class Contact(
     @Column(updatable = false)
     val partitionAreaId: Long = 0L,
 
-    ) : BaseEntity()
+) : BaseEntity()
 
 @Immutable
 @Entity
