+++ projects/workforce-allocations-to-delius/src/main/kotlin/uk/gov/justice/digital/hmpps/integrations/delius/contact/ContactTypeNotFoundException.kt
@@ -1,4 +1,3 @@
 package uk.gov.justice.digital.hmpps.integrations.delius.contact
 
-class ContactTypeNotFoundException(val code: String) : RuntimeException("Contact Type not found: $code") {
-}
+class ContactTypeNotFoundException(val code: String) : RuntimeException("Contact Type not found: $code")
