+++ projects/workforce-allocations-to-delius/src/main/kotlin/uk/gov/justice/digital/hmpps/integrations/delius/event/Event.kt
@@ -1,4 +1,3 @@
 package uk.gov.justice.digital.hmpps.integrations.delius.event
 
-class Event {
-}
+class Event
