+++ projects/workforce-allocations-to-delius/src/main/kotlin/uk/gov/justice/digital/hmpps/integrations/delius/event/requirement/Requirement.kt
@@ -1,4 +1,3 @@
 package uk.gov.justice.digital.hmpps.integrations.delius.event.requirement
 
-class Requirement {
-}
+class Requirement
