+++ projects/workforce-allocations-to-delius/src/test/kotlin/uk/gov/justice/digital/hmpps/config/DeliusConnectionProviderTest.kt
@@ -7,7 +7,6 @@ import org.mockito.Mock
 import org.mockito.junit.jupiter.MockitoExtension
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
-import uk.gov.justice.digital.hmpps.data.generator.UserGenerator
 import uk.gov.justice.digital.hmpps.integrations.delius.audit.service.UserService
 import java.sql.Connection
 import java.sql.PreparedStatement
