+++ projects/workforce-allocations-to-delius/src/test/kotlin/uk/gov/justice/digital/hmpps/integrations/delius/audit/AuditorAwareConfigurationTest.kt
@@ -12,7 +12,6 @@ import org.springframework.security.core.Authentication
 import org.springframework.security.core.context.SecurityContext
 import org.springframework.security.core.context.SecurityContextHolder
 import uk.gov.justice.digital.hmpps.config.security.ServicePrincipal
-import uk.gov.justice.digital.hmpps.data.generator.UserGenerator
 import uk.gov.justice.digital.hmpps.integrations.delius.audit.service.AuditorAware
 
 @ExtendWith(MockitoExtension::class)
